# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Pharmawatch API","description":"Pharmawatch API is a Solude's REST API that allows communication between the Pharmawatch web application and the Pharmawatch database on BigQuery to retrieve and upload data in a secure way and with the proper authentication.","version":"1.1.0"},"servers":[{"url":"https://pharmawatch-api-iqdmxo5f2a-rj.a.run.app/","description":"Homologation server"},{"url":"https://localhost:80/","description":"Docker local development server"},{"url":"http://localhost:8000/","description":"Local development server"}],"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/materiais/":{"get":{"tags":["Materials","Materials","Retrieval"],"summary":"Retrieve Complete ABC Curve of materials used in a given period","description":"Given a set of filters, retrieves the complete ABC Curve of materials used in a given period.","operationId":"get_materials_abc_curve_materiais__get","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_get_materials_abc_curve_materiais__get"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CurvaAbcMaterialResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/materiais/filtros":{"get":{"tags":["Materials","Materials","Filters","Retrieval"],"summary":"Retrieve Valid Materials Filters","description":"Retrieve the valid filters that can be applied to the materials list.","operationId":"get_medicamentos_filters_materiais_filtros_get","requestBody":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/UsuarioBase"}],"title":"User","default":{"token":"mocktoken","descr_prim_nome":"Mock","descr_ult_nome":"User","tags":[],"organizacao":99,"descr_email":"mockuser@mockorg.com","data_criacao":"30/11/2023"}}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FiltrosResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/upload/":{"post":{"tags":["Upload","Materials","Upload","Management"],"summary":"Upload Material Data","description":"Upload material data in CSV or Excel format to BigQuery.","operationId":"upload_material_data_upload__post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_material_data_upload__post"}}},"required":true},"responses":{"200":{"description":"Success message if upload is successful.","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/upload/reset-database":{"get":{"tags":["Upload","Removal","Admin"],"summary":"Reset Database","description":"Reset the database to its initial state.","operationId":"reset_database_upload_reset_database_get","requestBody":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/UsuarioBase"}],"title":"User","default":{"token":"mocktoken","descr_prim_nome":"Mock","descr_ult_nome":"User","tags":[],"organizacao":99,"descr_email":"mockuser@mockorg.com","data_criacao":"30/11/2023"}}}}},"responses":{"200":{"description":"Success message if database is reset successfully.","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AbcBase":{"properties":{"material":{"type":"string","title":"Material"},"tipo":{"type":"string","title":"Tipo"},"quantidade_unidade":{"type":"number","title":"Quantidade Unidade"},"valor_unidade":{"type":"number","title":"Valor Unidade"},"valor_total":{"type":"number","title":"Valor Total"},"percentual_valor_total":{"type":"number","title":"Percentual Valor Total"},"curva_abc":{"type":"string","title":"Curva Abc"},"variacao_cp":{"type":"number","title":"Variacao Cp"}},"type":"object","required":["material","tipo","quantidade_unidade","valor_unidade","valor_total","percentual_valor_total","curva_abc","variacao_cp"],"title":"AbcBase"},"Body_get_materials_abc_curve_materiais__get":{"properties":{"filters":{"$ref":"#/components/schemas/FiltrosMaterialRequest"},"user":{"allOf":[{"$ref":"#/components/schemas/UsuarioBase"}],"default":{"data_criacao":"30/11/2023","descr_email":"mockuser@mockorg.com","descr_prim_nome":"Mock","descr_ult_nome":"User","organizacao":99,"tags":[],"token":"mocktoken"}}},"type":"object","required":["filters"],"title":"Body_get_materials_abc_curve_materiais__get"},"Body_upload_material_data_upload__post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"user":{"allOf":[{"$ref":"#/components/schemas/UsuarioBase"}],"default":{"data_criacao":"30/11/2023","descr_email":"mockuser@mockorg.com","descr_prim_nome":"Mock","descr_ult_nome":"User","organizacao":99,"tags":[],"token":"mocktoken"}}},"type":"object","required":["file"],"title":"Body_upload_material_data_upload__post"},"CurvaAbcMaterialResponse":{"properties":{"page":{"type":"integer","exclusiveMinimum":0.0,"title":"Page","default":1},"per_page":{"type":"integer","exclusiveMinimum":0.0,"title":"Per Page","default":10},"data":{"items":{"$ref":"#/components/schemas/AbcBase"},"type":"array","title":"Data"},"total_pages":{"type":"integer","title":"Total Pages"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["data","total_pages","total"],"title":"CurvaAbcMaterialResponse"},"DestinoBase":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"descr_destino":{"type":"string","title":"Descr Destino"}},"type":"object","required":["descr_destino"],"title":"DestinoBase"},"FiltrosMaterialRequest":{"properties":{"page":{"type":"integer","exclusiveMinimum":0.0,"title":"Page","default":1},"per_page":{"type":"integer","exclusiveMinimum":0.0,"title":"Per Page","default":10},"ordenacao":{"items":{"anyOf":[{"type":"string"},{"type":"null"}]},"type":"array","title":"Ordenacao"},"ordenacao_crescente":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Ordenacao Crescente"},"id_material":{"items":{"anyOf":[{"type":"integer"},{"type":"null"}]},"type":"array","title":"Id Material"},"id_tipo":{"items":{"anyOf":[{"type":"integer"},{"type":"null"}]},"type":"array","title":"Id Tipo"},"id_origem":{"items":{"anyOf":[{"type":"integer"},{"type":"null"}]},"type":"array","title":"Id Origem"},"id_destino":{"items":{"anyOf":[{"type":"integer"},{"type":"null"}]},"type":"array","title":"Id Destino"},"data_inicio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Data Inicio"},"data_fim":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Data Fim"}},"type":"object","title":"FiltrosMaterialRequest"},"FiltrosResponse":{"properties":{"material":{"items":{"$ref":"#/components/schemas/MaterialBase"},"type":"array","title":"Material"},"tipo":{"items":{"$ref":"#/components/schemas/TipoBase"},"type":"array","title":"Tipo"},"origem":{"items":{"$ref":"#/components/schemas/OrigemBase"},"type":"array","title":"Origem"},"destino":{"items":{"$ref":"#/components/schemas/DestinoBase"},"type":"array","title":"Destino"}},"type":"object","required":["material","tipo","origem","destino"],"title":"FiltrosResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MaterialBase":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"descr_material":{"type":"string","title":"Descr Material"}},"type":"object","required":["descr_material"],"title":"MaterialBase"},"OrigemBase":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"descr_origem":{"type":"string","title":"Descr Origem"}},"type":"object","required":["descr_origem"],"title":"OrigemBase"},"TipoBase":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"descr_tipo":{"type":"string","title":"Descr Tipo"}},"type":"object","required":["descr_tipo"],"title":"TipoBase"},"UsuarioBase":{"properties":{"id_usuario":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id Usuario"},"token":{"type":"string","title":"Token"},"descr_prim_nome":{"type":"string","title":"Descr Prim Nome"},"descr_ult_nome":{"type":"string","title":"Descr Ult Nome"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]},"organizacao":{"type":"integer","title":"Organizacao"},"descr_email":{"type":"string","title":"Descr Email"},"data_criacao":{"type":"string","title":"Data Criacao"}},"type":"object","required":["token","descr_prim_nome","descr_ult_nome","organizacao","descr_email","data_criacao"],"title":"UsuarioBase"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"Filters","description":"Operations related to filters of data"},{"name":"Materials","description":"Operations related to hospital materials"},{"name":"Upload","description":"Operations related to data upload of data"},{"name":"Retrieval","description":"Operations related to data retrieval of data"},{"name":"Management","description":"Operations related to data management as upload and removal in user level"},{"name":"Removal","description":"Operations related to data removal"},{"name":"Admin","description":"Operations related to administrative tasks"}]}